src/main/java/tr/edu/medipol/yova/birimtestornek1/Hesap.java
package tr.edu.medipol.yova.birimtestornek1;

public class Hesap {
	
	public static int topla(int sayi1, int sayi2) {
		
		return sayi1 + sayi2;
	}
	

	public static int cikar(int sayi1, int sayi2) {
		
		return sayi1 - sayi2;
		}
	
	
	public static int carp(int sayi1, int sayi2) {
	
	return sayi1 * sayi2;
	
	}
		
		
	public static int bol(int sayi1, int sayi2) {
	
		 if (sayi2 == 0) {
		        throw new IllegalArgumentException("Bölen sıfır olamaz!");
		    }
		    return sayi1 / sayi2;

	}
}
